verenc/cd00: Move `encrypt` to package probenc
